refactor(test): align admin-settings tests with per-feature modules
Mirror the settings-*.ts source split on the test side so every settings
test file is named after the feature/route it covers and lives under
test/lib/server-settings/.

Moves (rename only):
- server-wallet.test.ts          -> server-settings/apple-wallet.test.ts
- server-google-wallet.test.ts   -> server-settings/google-wallet.test.ts
- email-templates-admin.test.ts  -> server-settings/email-templates.test.ts
- server-attendee-statuses.test.ts -> server-settings/statuses.test.ts
- header-image.test.ts           -> server-settings/header-image.test.ts

Catch-all breakups (sources removed):
- server-settings-advanced.test.ts -> email, show-public-api (new), plus
  reset-database / settings-page / domains merges
- server-settings-advanced-custom-domain.test.ts -> domains (new),
  column-order (new), settings-page

Dedupe:
- merged server-embed-hosts.test.ts into server-settings/embed-hosts.test.ts

Pure test relocation, no assertions changed (only exact embed-hosts
duplicates dropped). precommit fully green: coverage stays at 100%.
