+++ test/lib/server-settings/email-templates.test.ts
